Minor refactor
* East const everywhere
* Use forward declarations where applicable
* Move some functions
* Hide some function implementations
* Update CMake module dependencies
