+++ common/src/header_check.cc
@@ -1,5 +1,4 @@
 #include "handle_wrapper.hh"
 #include "to_array.hh"
 #include "vulkan_helper_types.hh"
-#include "vulkan_to_string.hh"
 #include "vulkan_version.hh"
