+++ glfw_wrappers/include/glfw_window_wrapper.hh
@@ -1,8 +1,9 @@
 #ifndef GLFW_WINDOW_WRAPPER_HH_INCLUDED_
 #define GLFW_WINDOW_WRAPPER_HH_INCLUDED_
 
-#include "GLFW/glfw3.h"
 #include "handle_wrapper.hh"
+
+class GLFWwindow;
 
 namespace glfw {
