+++ glfw_wrappers/src/glfw_window_wrapper.cc
@@ -2,6 +2,8 @@
 
 #include "fmt/core.h"
 
+#include <GLFW/glfw3.h>
+
 namespace glfw {
 
 auto
