+++ vulkan_wrappers/src/vulkan_logical_device_wrapper.cc
@@ -1,6 +1,7 @@
 #include "vulkan_logical_device_wrapper.hh"
 
 #include "fmt/core.h"
+#include "vulkan_helper_types.hh"
 
 #include <set>
 
@@ -20,8 +21,8 @@ LogicalDeviceWrapper::errorMsg(const std::string & msg) -> void
 
 auto
 LogicalDeviceWrapper::create(VkPhysicalDevice              phys_dev,
-                             const QueueFamilyIndices &    indices,
-                             gsl::span<const char * const> extensions)
+                             QueueFamilyIndices const &    indices,
+                             gsl::span<char const * const> extensions)
     -> std::optional<LogicalDeviceWrapper>
 {
     if (!indices.graphics_family.has_value()) {
@@ -29,14 +30,14 @@ LogicalDeviceWrapper::create(VkPhysicalDevice              phys_dev,
         return std::nullopt;
     }
 
-    const std::set<uint32_t> unique_queue_families = {*indices.graphics_family,
+    std::set<uint32_t> const unique_queue_families = {*indices.graphics_family,
                                                       *indices.present_family};
 
     std::vector<VkDeviceQueueCreateInfo> queue_create_infos{};
     queue_create_infos.reserve(unique_queue_families.size());
 
     constexpr float queue_priority = 1.0F;
-    for (const uint32_t queue_family : unique_queue_families) {
+    for (uint32_t const queue_family : unique_queue_families) {
         queue_create_infos.emplace_back();
         auto & info = queue_create_infos.back();
         info.sType = VK_STRUCTURE_TYPE_DEVICE_QUEUE_CREATE_INFO;
