+++ vulkan_wrappers/src/vulkan_surface_wrapper.cc
@@ -1,7 +1,10 @@
 #include "vulkan_surface_wrapper.hh"
 
+#include "vulkan_instance_wrapper.hh"
 #include "fmt/core.h"
 #include "glfw_window_wrapper.hh"
+
+#include <GLFW/glfw3.h>
 
 namespace vulkan {
